Update FFA test application to work with UnitTestLib (#27)
## Description

This change updated the test app to work with standard UEFI unit test
framework and will produce properly parsable test results for pipeline
to report.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This change is tested by running the test application on QEMU SBSA
platform and passed in UEFI shell environment.

## Integration Instructions

N/A