Fixing TPM related modules (#31)
## Description

As BASECORE and SILICON ARM repo updates to follow the support from
EDK2, this change is created to update the file paths to fix the
pipeline builds.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [ ] Impacts functionality?
- [ ] Impacts security?
- [x] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This is tested on QEMU SBSA and booted to UEFI shell.

## Integration Instructions

N/A