Fixing FFA notification unit test (#33)
## Description

The current test application does not really check whether the
notification set command is issued from test services.

This change adds a global boolean variable to be cleared from the
interrupt handler.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

The test still passes after the change.

## Integration Instructions

N/A