[Rebase & FF] Adding 2 services to Rust based secure partition using ODP framework (#35)
## Description

This change adds the foundation work for Rust based secure partition.

It currently supports Test service and notify service from
`haf-ec-service` crate.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [x] Impacts functionality?
- [x] Impacts security?
- [x] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This was tested with the same FFA functionality test and passed all unit
tests.

## Integration Instructions

N/A

---------

Signed-off-by: Kun Qin <kun.qin@microsoft.com>