FfaFeaturePkg/SecurePartitionEntryPoint: Add gHobList
The MsSecurePartition.inf module is of type `MM_CORE_STANDALONE`.

Although it does not directly depend on `HobLib`, libraries attached
to it could. The `MM_CORE_STANDALONE` instance of `HobLib` in
StandaloneMmPkg depends on the symbol `gHobList` to be defined in
the core. Add this symbol to `SecurePartitionEntryPoint` which is
equivalent to where the symbol is defined for the Arm Standalone
MM Core:

  `ArmPkg\/ibrary/ArmStandaloneMmCoreEntryPoint`

This change adds that symbol so the library instance can successfully
be linked.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>