FfaFeaturePkg/SecurePartitionEntryPoint: Add gHobList (#45)
## Description

The MsSecurePartition.inf module is of type `MM_CORE_STANDALONE`.

Although it does not directly depend on `HobLib`, libraries attached
to it could. The `MM_CORE_STANDALONE` instance of `HobLib` in
StandaloneMmPkg depends on the symbol `gHobList` to be defined in
the core. Add this symbol to `SecurePartitionEntryPoint` which is
equivalent to where the symbol is defined for the Arm Standalone
MM Core:

  `ArmPkg/Library/ArmStandaloneMmCoreEntryPoint`

This change adds that symbol so HOB the library instance can
successfully be linked.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

- FfaFeaturePkg build
- Integration of MsSecurePartition into a platform

## Integration Instructions

- The platform should ensure a `MM_CORE_STANDALONE` instance of `HobLib`
is defined in the platform DSC.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>