Build rust secure partitions when releases are made (#46)
## Description

This change brings the pipeline building of rust secure partitions to be
tied to releases of this repo.

Instead of relying on the secure partitions to be built as part of the
build process of platform, this change took the approach of releasing
binary files to be consumed by platforms.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This was tested on pipeline build and ensure the binaries are properly
generated.

## Integration Instructions

One should ingest the artifacts associated with a particular release if
the secure partition is to be integrated in the platform level.