Fix a path issue in the git artifact upload step (#51)
## Description

The previous fix missed a step to update the path where the zipped
artifacts should be uploaded.

This change is created to fix this issue.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

The change matches the previous step that creates the binaries.

## Integration Instructions

N/A