Simplify release attachment process for secure partitions (#53)
## Description

This change removes the mangling step of uploading and downloading
artifacts to simplify the file staging process.

And this also fixes an issue where the original script will not be able
to locate the downloaded artifacts.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

The build pipeline passes during dry runs.

## Integration Instructions

N/A