[FF-A] Add FFA_NS_RES_INFO_GET Command (#57)
## Description

Add FFA_NS_RES_INFO_GET function call to the ARM FF-A library. Add
relevant structures to the header. Update FfaPartitionTestApp to include
test for FFA_NS_RES_INFO_GET.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested and verified through the QemuSbsa build with the
FfaPartitionTestApp.

## Integration Instructions

N/A